Move QuantizedMeshLoader to CesiumLegacyTerrain.
