+++ Cesium3DTilesSelection/src/LayerJsonTerrainLoader.cpp
@@ -1,10 +1,10 @@
 #include "LayerJsonTerrainLoader.h"
 
-#include <Cesium3DTilesContent/QuantizedMeshLoader.h>
 #include <Cesium3DTilesContent/upsampleGltfForRasterOverlays.h>
 #include <CesiumAsync/IAssetResponse.h>
 #include <CesiumGeospatial/calcQuadtreeMaxGeometricError.h>
 #include <CesiumGltfContent/GltfUtilities.h>
+#include <CesiumLegacyTerrain/QuantizedMeshLoader.h>
 #include <CesiumRasterOverlays/RasterOverlayUtilities.h>
 #include <CesiumUtility/JsonHelpers.h>
 #include <CesiumUtility/Uri.h>
@@ -17,6 +17,7 @@ using namespace Cesium3DTilesContent;
 using namespace Cesium3DTilesSelection;
 using namespace CesiumGeometry;
 using namespace CesiumGeospatial;
+using namespace CesiumLegacyTerrain;
 using namespace CesiumRasterOverlays;
 using namespace CesiumUtility;
