+++ Cesium3DTilesWriter/src/SubtreeWriter.cpp
@@ -1,7 +1,7 @@
 #include "Cesium3DTilesWriter/SubtreeWriter.h"
 
 #include "TilesetJsonWriter.h"
-#include "registerExtensions.h"
+#include "registerWriterExtensions.h"
 
 #include <CesiumJsonWriter/JsonWriter.h>
 #include <CesiumJsonWriter/PrettyJsonWriter.h>
@@ -9,7 +9,7 @@
 
 namespace Cesium3DTilesWriter {
 
-SubtreeWriter::SubtreeWriter() { registerExtensions(this->_context); }
+SubtreeWriter::SubtreeWriter() { registerWriterExtensions(this->_context); }
 
 CesiumJsonWriter::ExtensionWriterContext& SubtreeWriter::getExtensions() {
   return this->_context;
