+++ Cesium3DTilesWriter/src/TilesetWriter.cpp
@@ -1,7 +1,7 @@
 #include "Cesium3DTilesWriter/TilesetWriter.h"
 
 #include "TilesetJsonWriter.h"
-#include "registerExtensions.h"
+#include "registerWriterExtensions.h"
 
 #include <CesiumJsonWriter/JsonWriter.h>
 #include <CesiumJsonWriter/PrettyJsonWriter.h>
@@ -9,7 +9,7 @@
 
 namespace Cesium3DTilesWriter {
 
-TilesetWriter::TilesetWriter() { registerExtensions(this->_context); }
+TilesetWriter::TilesetWriter() { registerWriterExtensions(this->_context); }
 
 CesiumJsonWriter::ExtensionWriterContext& TilesetWriter::getExtensions() {
   return this->_context;
